services/gatway-api/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.1
nats-py==2.7.2
prometheus-client==0.20.0